Use size_t as array/vector index

+++ worker/include/RTC/RtpDictionaries.h
@@ -281,7 +281,7 @@ namespace RTC
 
 		Json::Value toJson();
 		void ReduceHeaderExtensions(std::vector<RtpHeaderExtension>& supportedHeaderExtensions);
-		uint32_t GetClockRateForEncoding(uint8_t encodingIdx);
+		uint32_t GetClockRateForEncoding(size_t encodingIdx);
 
 	private:
 		void ValidateCodecs();
