+++ worker/src/RTC/RtpDictionaries/RtpParameters.cpp
@@ -184,12 +184,12 @@ namespace RTC
 		this->headerExtensions = updatedHeaderExtensions;
 	}
 
-	uint32_t RtpParameters::GetClockRateForEncoding(uint8_t encodingIdx)
+	uint32_t RtpParameters::GetClockRateForEncoding(size_t encodingIdx)
 	{
 		MS_TRACE();
 
 		if (this->encodings.size() < encodingIdx + 1)
-			MS_ABORT("no such a encoding [encodingIdx:%" PRIu8 "]", encodingIdx);
+			MS_ABORT("no such a encoding [encodingIdx:%zu]", encodingIdx);
 
 		uint8_t payloadType = this->encodings[encodingIdx].codecPayloadType;
 		uint32_t clockRate = 0;
@@ -210,7 +210,7 @@ namespace RTC
 		// This should never happen.
 		if (it == this->codecs.end())
 		{
-			MS_ABORT("no valid codec payload type for the requested encoding [encodingIdx:%" PRIu8 "]", encodingIdx);
+			MS_ABORT("no valid codec payload type for the requested encoding [encodingIdx:%zu]", encodingIdx);
 		}
 
 		return clockRate;
